New Development: No             JIRA Issue: Yes CAS-5015,5016,5018,5019         Ready for Test: Yes      Interface Changes: Yes What Interface Changed: For more accuracy changed some method parameters from float to double.          Test Programs: List test programs   Put in Release Notes: No              Module(s): Module Names change impacts.            Description: Viewer Search Functionality Bugs from Leroy.
git-svn-id: https://svn.cv.nrao.edu/svn/casa/branches/release-4_1_0@23788 ac9c0da5-131f-0410-9009-a9c195f183e6

+++ code/display/QtPlotter/MomentSettingsWidgetOptical.cc
@@ -149,7 +149,7 @@ void MomentSettingsWidgetOptical::setUnits( QString unitStr ){
 	ui.collapseUnits->setText( unitStr );
 }
 
-void MomentSettingsWidgetOptical::setRange( float xmin, float xmax ){
+void MomentSettingsWidgetOptical::setRange( double xmin, double xmax ){
 	if (xmax < xmin){
 		ui.startValue->clear();
 		ui.endValue->clear();
